Fix test value range of FullDistortionKernel32Bits
Currently the unit tests use unsigned input data up to (1 << 15).
This makes it more difficult to catch overflow errors when using
non-widening multiplication operations. In real encoding scenarios
values can exceed (1 << 15), in the test vectors for end-to-end tests
values are found to be between -(1 << 16) and (1 << 17). This change
makes the FullDistortionKernel32Bits test generate input data between
-(1 << 17) and (1 << 17) to improve test coverage.