api: add option to set a logger
This is a very basic logging callback with an optional context.
This does not replicate what FFmpeg's libavutil has, as that one is
a lot more complex and even passes around va_list, as opposed to our
simple pre-formatted string approach. However, this should at least
allow customizing and interacting with the log outputs to some extent
without needing to handle file IO.

Signed-off-by: Christopher Degawa <1480c1@videolan.org>